refactor(gateway): extract MtlsConfig (GW-2 S6)
Move MtlsConfig struct, its 10 default_mtls_* helpers, its impl Default,
and header-defaults test. A small test_default_mtls_headers now lives in
mtls.rs (unit-level) while test_default_mtls_disabled stays in config.rs
(wiring into Config::default).
